+++ spec/models/renalware/accesses/catheter_insertion_technique_spec.rb
@@ -1,0 +1,10 @@
+require "rails_helper"
+
+module Renalware
+  module Accesses
+    describe CatheterInsertionTechnique do
+      it { is_expected.to validate_presence_of(:code) }
+      it { is_expected.to validate_presence_of(:description) }
+    end
+  end
+end